requirements.txt
numpy~=1.26.4
tensorflow~=2.18.0
ai-edge-litert==1.0.1
opencv-python==4.10.0.84
supervision==0.25.0
paho-mqtt==2.1.0
pillow==11.0.0
fonttools==4.55.0